refactor: move the `defineProjectConfig` function to its own file
